Are: fixed warning in MusicClient::send function by casting to unsigned int. Removed executables from git.

+++ MusicClient.cpp
@@ -74,7 +74,7 @@ int main(int argc, char *argv[]) {
       ph.type = 0;
       ph.length = 0;
       unsigned long bufferLen = serializePacket(buffer, ph);
-      if (send(sock, buffer, bufferLen, 0) != bufferLen)
+      if (send(sock, buffer, bufferLen, 0) != (unsigned int)bufferLen)
         DieWithError("send() sent a different number of bytes than expected");
 
       recvTCPMessage(sock, buffer, recv_buffer);
